+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # CHANGELOG
 
+## 0.17.1
+- fix: invalid function call
+
 ## 0.17.0
 - feat: upgrade webpack-dev-server: 4.2.1 -> 4.3.0
 - chore: refactor & optimise